code rearranged
