+++ icicle/backend/cpu/include/cpu_ntt_main.h
@@ -1,7 +1,6 @@
 #pragma once
-#include "icicle/errors.h"
-#include "icicle/ntt.h"
-#include "ntt_tasks.h"
+#include "icicle/utils/log.h"
+#include "ntt_cpu.h"
 #include "ntt_tasks_ref.h"
 #include <iostream>
 
@@ -42,77 +41,14 @@ namespace ntt_cpu {
   eIcicleError
   cpu_ntt(const Device& device, const E* input, uint64_t size, NTTDir direction, const NTTConfig<S>& config, E* output)
   {
-    const uint32_t domain_max_size = CpuNttDomain<S>::s_ntt_domain.get_max_size(); 
     ICICLE_ASSERT(!(size & (size - 1))) << "Size must be a power of 2. size = " << size;
     ICICLE_ASSERT(size <= CpuNttDomain<S>::s_ntt_domain.get_max_size()) << "Size is too large for domain. size = " << size << ", domain_max_size = " << CpuNttDomain<S>::s_ntt_domain.get_max_size();
 
-    const uint32_t logn = uint32_t(log2(size));
-    const uint64_t total_input_size = size * config.batch_size;
-    
-    NttCpu<S, E> ntt(logn, direction, config);
-    NttTaskCordinates ntt_task_cordinates = {0, 0, 0, 0, 0};
-    NttTasksManager<S, E> ntt_tasks_manager(logn);
-    const uint32_t nof_threads = std::thread::hardware_concurrency();
-    auto tasks_manager = new TasksManager<NttTask<S, E>>(nof_threads - 1);
-    NttTask<S, E>* task_slot;
-    std::unique_ptr<S[]> arbitrary_coset = nullptr;
-    const uint32_t coset_stride = ntt.find_or_generate_coset(arbitrary_coset);
+    NttCpu<S, E> ntt(uint32_t(log2(size)), direction, config, input, output);
+    ntt.run();
 
-    ntt.copy_and_reorder_if_needed(input, output);
-    if (config.coset_gen != S::one() && direction == NTTDir::kForward) {
-      ntt.coset_mul(output, coset_stride, arbitrary_coset);
-    }
-
-    if (logn > HIERARCHY_1) {
-      for (ntt_task_cordinates.hierarchy_1_layer_idx = 0; ntt_task_cordinates.hierarchy_1_layer_idx < 2; ntt_task_cordinates.hierarchy_1_layer_idx++) {
-        const uint32_t sunbtt_plus_batch_logn = ntt.ntt_sub_logn.hierarchy_1_layers_sub_logn[ntt_task_cordinates.hierarchy_1_layer_idx] + uint32_t(log2(config.batch_size));
-        const uint32_t log_nof_hierarchy_1_subntts_todo_in_parallel = sunbtt_plus_batch_logn < HIERARCHY_1 ? HIERARCHY_1 - sunbtt_plus_batch_logn : 0;
-        const uint32_t nof_hierarchy_1_subntts_todo_in_parallel = 1 << log_nof_hierarchy_1_subntts_todo_in_parallel;
-        const uint32_t log_nof_subntts_chunks = ntt.ntt_sub_logn.hierarchy_1_layers_sub_logn[1 - ntt_task_cordinates.hierarchy_1_layer_idx] -log_nof_hierarchy_1_subntts_todo_in_parallel;
-        const uint32_t nof_subntts_chunks = 1 << log_nof_subntts_chunks;
-        for (uint32_t hierarchy_1_subntts_chunck_idx = 0; hierarchy_1_subntts_chunck_idx < nof_subntts_chunks; hierarchy_1_subntts_chunck_idx++) {
-          for (uint32_t hierarchy_1_subntt_idx_in_chunck = 0; hierarchy_1_subntt_idx_in_chunck < nof_hierarchy_1_subntts_todo_in_parallel; hierarchy_1_subntt_idx_in_chunck++) {
-            ntt_task_cordinates.hierarchy_1_subntt_idx = hierarchy_1_subntts_chunck_idx * nof_hierarchy_1_subntts_todo_in_parallel + hierarchy_1_subntt_idx_in_chunck;
-            ntt.hierarchy1_push_tasks(output, ntt_task_cordinates, ntt_tasks_manager);
-          }
-          ntt.handle_pushed_tasks(tasks_manager, ntt_tasks_manager, ntt_task_cordinates.hierarchy_1_layer_idx);
-        }
-        if (ntt_task_cordinates.hierarchy_1_layer_idx == 0) { ntt.hierarchy_1_reorder(output); }
-      }
-      // reset hierarchy_1_subntt_idx so that reorder_and_refactor_if_needed will calculate the correct memory index
-      ntt_task_cordinates.hierarchy_1_subntt_idx = 0;
-      if (config.columns_batch) {
-        ntt.reorder_and_refactor_if_needed(output, ntt_task_cordinates, true);
-      } else {
-        for (uint32_t b = 0; b < config.batch_size; b++) {
-          ntt.reorder_and_refactor_if_needed(output + b * size, ntt_task_cordinates, true);
-        }
-      }
-    } else {
-      ntt.hierarchy1_push_tasks(output, ntt_task_cordinates, ntt_tasks_manager);
-      ntt.handle_pushed_tasks(tasks_manager, ntt_tasks_manager, 0);
-    }
-
-    
-    if (direction == NTTDir::kInverse) {
-      S inv_size = S::inv_log_size(logn);
-      for (uint64_t i = 0; i < total_input_size; ++i) {
-        output[i] = output[i] * inv_size;
-      }
-      if (config.coset_gen != S::one()) { ntt.coset_mul(output, coset_stride, arbitrary_coset); }
-    }
-
-    if (config.ordering == Ordering::kNR || config.ordering == Ordering::kRR) {
-      ntt_task_cordinates = {0, 0, 0, 0, 0};
-      ntt.reorder_by_bit_reverse(ntt_task_cordinates, output, true);
-    }
-    delete tasks_manager;
     return eIcicleError::SUCCESS;
   }
-
-
-
-
 
   template <typename S = scalar_t, typename E = scalar_t>
   eIcicleError
@@ -152,14 +88,14 @@ namespace ntt_cpu {
     uint32_t log_nof_subntts_chunks;
     uint32_t nof_subntts_chunks;
 
-    if (logn > HIERARCHY_1) {
+    if (logn > HIERARCHY_1_REF) {
       for (ntt_task_cordinates.hierarchy_1_layer_idx = 0; ntt_task_cordinates.hierarchy_1_layer_idx < 2;
            ntt_task_cordinates.hierarchy_1_layer_idx++) {
         sunbtt_plus_batch_logn =
           ntt.ntt_sub_logn.hierarchy_1_layers_sub_logn[ntt_task_cordinates.hierarchy_1_layer_idx] +
           uint32_t(log2(config.batch_size));
         log_nof_hierarchy_1_subntts_todo_in_parallel =
-          sunbtt_plus_batch_logn < HIERARCHY_1 ? HIERARCHY_1 - sunbtt_plus_batch_logn : 0;
+          sunbtt_plus_batch_logn < HIERARCHY_1_REF ? HIERARCHY_1_REF - sunbtt_plus_batch_logn : 0;
         nof_hierarchy_1_subntts_todo_in_parallel = 1 << log_nof_hierarchy_1_subntts_todo_in_parallel;
         log_nof_subntts_chunks =
           ntt.ntt_sub_logn.hierarchy_1_layers_sub_logn[1 - ntt_task_cordinates.hierarchy_1_layer_idx] -
@@ -192,6 +128,8 @@ namespace ntt_cpu {
       ntt.hierarchy1_push_tasks(output, ntt_task_cordinates, ntt_tasks_manager);
       ntt.handle_pushed_tasks(tasks_manager, ntt_tasks_manager, 0);
     }
+
+    // std::cout << "[REF] PRE NORMALIZE: right:\t["; for (int i = 0; i < total_input_size-1; i++) { std::cout << output[i] << ", "; } std::cout <<output[total_input_size-1]<<"]"<< std::endl;
 
     if (direction == NTTDir::kInverse) {
       S inv_size = S::inv_log_size(logn);
