+++ icicle/backend/cpu/include/ntt_utils.h
@@ -1,0 +1,50 @@
+#pragma once
+#include "icicle/fields/field_config.h"
+
+#include <_types/_uint32_t.h>
+#include <sys/types.h>
+
+using namespace field_config;
+using namespace icicle;
+namespace ntt_cpu {
+
+    /**
+    * @brief Represents the coordinates of a task in the NTT hierarchy.
+    * This struct holds indices that identify the position of a task within the NTT computation hierarchy.
+    *
+    * @param hierarchy_1_layer_idx Index of the hierarchy_1 layer.
+    * @param hierarchy_1_subntt_idx Index of the sub-NTT within the hierarchy_1 layer.
+    * @param hierarchy_0_layer_idx Index of the hierarchy_0 layer.
+    * @param hierarchy_0_block_idx Index of the block within the hierarchy_0 layer.
+    * @param hierarchy_0_subntt_idx Index of the sub-NTT within the hierarchy_0 block.
+    *
+    * @method bool operator==(const NttTaskCordinates& other) const Compares two task coordinates for equality.
+    */
+    struct NttTaskCordinates {
+        uint32_t hierarchy_1_layer_idx = 0;
+        uint32_t hierarchy_1_subntt_idx = 0;
+        uint32_t hierarchy_0_layer_idx = 0;
+        uint32_t hierarchy_0_block_idx = 0;
+        uint32_t hierarchy_0_subntt_idx = 0;
+        bool reorder = false;
+
+        bool operator==(const NttTaskCordinates& other) const
+        {
+        return hierarchy_1_layer_idx == other.hierarchy_1_layer_idx &&
+                hierarchy_1_subntt_idx == other.hierarchy_1_subntt_idx &&
+                hierarchy_0_layer_idx == other.hierarchy_0_layer_idx &&
+                hierarchy_0_block_idx == other.hierarchy_0_block_idx &&
+                hierarchy_0_subntt_idx == other.hierarchy_0_subntt_idx;
+        }
+    };
+
+    uint64_t bit_reverse(uint64_t i, uint32_t logn)
+    {
+        uint32_t rev = 0;
+        for (uint32_t j = 0; j < logn; ++j) {
+        if (i & (1 << j)) { rev |= 1 << (logn - 1 - j); }
+        }
+        return rev;
+    };
+
+} // namespace ntt_cpu
